+++ src/main/java/app/global/converter/JsonConverter.java
@@ -1,0 +1,40 @@
+package app.global.converter;
+
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import jakarta.persistence.AttributeConverter;
+import jakarta.persistence.Converter;
+import lombok.extern.slf4j.Slf4j;
+
+@Slf4j
+@Converter
+public class JsonConverter implements AttributeConverter<Object, String> {
+
+    private final ObjectMapper objectMapper = new ObjectMapper();
+
+    @Override
+    public String convertToDatabaseColumn(Object attribute) {
+        if (attribute == null) {
+            return null;
+        }
+        try {
+            return objectMapper.writeValueAsString(attribute);
+        } catch (JsonProcessingException e) {
+            log.error("JSON 변환 실패", e);
+            return null;
+        }
+    }
+
+    @Override
+    public Object convertToEntityAttribute(String dbData) {
+        if (dbData == null || dbData.isEmpty()) {
+            return null;
+        }
+        try {
+            return objectMapper.readValue(dbData, Object.class);
+        } catch (JsonProcessingException e) {
+            log.error("JSON 파싱 실패", e);
+            return null;
+        }
+    }
+}
